Refactor: Create function modify survey data
Created a helper function to handle the data processing to turn form
data into a form the server expects
