+++ assets/scripts/survey/events.js
@@ -6,6 +6,7 @@ const ui = require('./ui')
 const { apiUrl } = require('../config')
 const socket = require('socket.io-client')(apiUrl)
 const store = require('../store')
+const processSurveyData = require('./../../../lib/process-survey-data')
 
 const onIndexSurvey = event => {
   if (event) {
@@ -44,41 +45,7 @@ const onCreateSurvey = event => {
   event.preventDefault()
 
   const data = getFormFields(event.target)
-
-  const dataArray = []
-  for (const key in data) {
-    const obj = {}
-    obj[key] = data[key]
-    dataArray.push(obj)
-  }
-
-  const surveyPojo = {
-    survey: {
-      title: dataArray[0].survey.title,
-      questions: []
-    }
-  }
-
-  for (let i = 1; i < dataArray.length; i++) {
-    const question = dataArray[i][`question${i}`]
-    const questionPojo = {
-      inputType: 'radio', // hard coded for now, need to revisit
-      text: question.text,
-      index: i - 1,
-      responses: []
-    }
-
-    const optionsKeys = Object.keys(question).slice(1)
-    for (let i = 0; i < optionsKeys.length; i++) {
-      const optionPojo = {
-        value: question[optionsKeys[i]],
-        users: []
-      }
-      questionPojo.responses.push(optionPojo)
-    }
-
-    surveyPojo.survey.questions.push(questionPojo)
-  }
+  const surveyPojo = processSurveyData(data)
 
   api.createSurvey(surveyPojo)
     .then(ui.createSurveySuccess)
